EKS_nat_cdk/requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_eks